# csf/tools/codegen/generated/brainspine.openapi.yaml
...
{
  "openapi": "3.1.0",
  "info": {
    "title": "adapter",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/snapshot": {
      "get": {
        "tags": [
          "Research"
        ],
        "summary": "Show the latest measured work and health report.",
        "operationId": "Research_GetSnapshot",
        "responses": {
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/google.rpc.Status"
                }
              }
            }
          },
          "200": {
            "description": "A successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/candace.brainspine.v1.GetSnapshotResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/compile": {
      "post": {
        "tags": [
          "Research"
        ],
        "summary": "Prepare a driving recipe: check the recipe and turn it into steps the controller can run.",
        "operationId": "Research_Compile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "controller": {
                    "$ref": "#/components/schemas/candace.brainspine.v1.Controller"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/google.rpc.Status"
                }
              }
            }
          },
          "200": {
            "description": "A successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/candace.brainspine.v1.CompileResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "candace.brainspine.v1.ArtifactEvidence": {
        "type": "object",
        "properties": {
          "bytes": {
            "type": "string",
            "format": "uint64"
          },
          "missing": {
            "type": "boolean"
          },
          "path": {
            "type": "string"
          },
          "sha256": {
            "type": "string"
          }
        }
      },
...
      "candace.brainspine.v1.CommandReceipt": {
        "type": "object",
        "description": "Command execution evidence is generated into every consumer. Success means\nthe recorded command exited zero; it does not prove an arbitrary claim.",
        "properties": {
          "branch": {
            "type": "string"
          },
          "claim": {
            "type": "string"
          },
          "command": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "cwd": {
            "type": "string"
          },
          "dashboardUrl": {
            "type": "string"
          },
          "durationSeconds": {
            "type": "number",
            "format": "double"
          },
          "evidence": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/candace.brainspine.v1.ArtifactEvidence"
            }
          },
          "exitCode": {
            "type": "integer",
            "format": "int32"
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "receiptId": {
            "type": "string"
          },
          "revision": {
            "type": "string"
          },
          "schemaVersion": {
            "type": "integer",
            "format": "int64"
          },
          "scope": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "taskId": {
            "type": "string"
          },
          "ticketUrl": {
            "type": "string"
          },
          "trackedChanges": {
            "type": "boolean"
          },
          "turnId": {
            "type": "string"
          }
        }
      },
      "candace.brainspine.v1.CompileResponse": {
        "type": "object",
        "properties": {
          "program": {
            "$ref": "#/components/schemas/candace.brainspine.v1.Program"
          }
        }
      },
      "candace.brainspine.v1.Controller": {
        "type": "object",
        "properties": {
          "acceleration": {
            "$ref": "#/components/schemas/candace.brainspine.v1.Expression"
          },
          "name": {
            "type": "string"
          },
          "schemaVersion": {
            "type": "integer",
            "format": "int64"
          },
          "steering": {
            "$ref": "#/components/schemas/candace.brainspine.v1.Expression"
          }
        }
      },
...
      "candace.brainspine.v1.Expression": {
        "type": "object",
        "properties": {
          "arguments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/candace.brainspine.v1.Expression"
            }
          },
          "inputIndex": {
            "type": "integer",
            "format": "int64"
          },
          "lower": {
            "type": "string",
            "format": "int64"
          },
          "opcode": {
            "$ref": "#/components/schemas/candace.brainspine.v1.Opcode"
          },
          "upper": {
            "type": "string",
            "format": "int64"
          },
          "value": {
            "type": "string",
            "format": "int64"
          }
        }
      },
...
      "candace.brainspine.v1.GetSnapshotResponse": {
        "type": "object",
        "properties": {
          "snapshot": {
            "$ref": "#/components/schemas/candace.brainspine.v1.Snapshot"
          }
        }
      },
...
      "candace.brainspine.v1.Instruction": {
        "type": "object",
        "properties": {
          "inputIndex": {
            "type": "integer",
            "format": "int64"
          },
          "lower": {
            "type": "string",
            "format": "int64"
          },
          "opcode": {
            "$ref": "#/components/schemas/candace.brainspine.v1.Opcode"
          },
          "upper": {
            "type": "string",
            "format": "int64"
          },
          "value": {
            "type": "string",
            "format": "int64"
          }
        }
      },
...
      "candace.brainspine.v1.Measurement": {
        "type": "object",
        "properties": {
          "candidateId": {
            "type": "string"
          },
          "metric": {
            "type": "string"
          },
          "runId": {
            "type": "string"
          },
          "split": {
            "type": "string"
          },
          "step": {
            "type": "string",
            "format": "uint64"
          },
          "value": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "candace.brainspine.v1.MetricDefinition": {
        "type": "object",
        "description": "Measurements are registered as data before their first observation. Metric\nnames and units stay stable across runs; hashes and IDs are receipt fields,\nnot unbounded Prometheus label dimensions. Missing data is never zero.",
        "properties": {
          "description": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          }
        }
      },
      "candace.brainspine.v1.Opcode": {
        "type": "string",
        "description": "All expression values are dimensionless fixed-point scalars with scale 1000.\nThe simulator adapter owns physical-unit normalization; actuator conversion\nis fixed by the spine, never supplied by a candidate. Every operation saturates\nto [-1000000000, 1000000000]. SCALE divides by 1000 toward zero.",
        "enum": [
          "OPCODE_UNSPECIFIED",
          "OPCODE_CONSTANT",
          "OPCODE_INPUT",
          "OPCODE_ADD",
          "OPCODE_SCALE",
          "OPCODE_CLAMP"
        ]
      },
...
      "candace.brainspine.v1.Program": {
        "type": "object",
        "properties": {
          "acceleration": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/candace.brainspine.v1.Instruction"
            }
          },
          "controllerHash": {
            "type": "string"
          },
          "schemaVersion": {
            "type": "integer",
            "format": "int64"
          },
          "steering": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/candace.brainspine.v1.Instruction"
            }
          }
        }
      },
...
      "candace.brainspine.v1.ResearchEvent": {
        "type": "object",
        "description": "Exactly one payload is present. Consumers validate the metric registry,\nfinite values and receipt references before displaying observations.",
        "properties": {
          "definition": {
            "$ref": "#/components/schemas/candace.brainspine.v1.MetricDefinition"
          },
          "measurement": {
            "$ref": "#/components/schemas/candace.brainspine.v1.Measurement"
          },
          "receipt": {
            "$ref": "#/components/schemas/candace.brainspine.v1.CommandReceipt"
          },
          "recordedAt": {
            "type": "string"
          },
          "schemaVersion": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "$ref": "#/components/schemas/candace.brainspine.v1.RunStatus"
          }
        },
        "oneOf": [
          {
            "type": "object",
            "not": {
              "anyOf": [
                {
                  "required": [
                    "definition"
                  ]
                },
                {
                  "required": [
                    "measurement"
                  ]
                },
                {
                  "required": [
                    "status"
                  ]
                },
                {
                  "required": [
                    "receipt"
                  ]
                }
              ]
            }
          },
          {
            "type": "object",
            "properties": {
              "definition": {
                "$ref": "#/components/schemas/candace.brainspine.v1.MetricDefinition"
              }
            },
            "required": [
              "definition"
            ]
          },
          {
            "type": "object",
            "properties": {
              "measurement": {
                "$ref": "#/components/schemas/candace.brainspine.v1.Measurement"
              }
            },
            "required": [
              "measurement"
            ]
          },
          {
            "type": "object",
            "properties": {
              "status": {
                "$ref": "#/components/schemas/candace.brainspine.v1.RunStatus"
              }
            },
            "required": [
              "status"
            ]
          },
          {
            "type": "object",
            "properties": {
              "receipt": {
                "$ref": "#/components/schemas/candace.brainspine.v1.CommandReceipt"
              }
            },
            "required": [
              "receipt"
            ]
          }
        ]
      },
      "candace.brainspine.v1.RunStatus": {
        "type": "object",
        "properties": {
          "evidencePath": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "phase": {
            "type": "string"
          },
          "runId": {
            "type": "string"
          }
        }
      },
...
      "candace.brainspine.v1.Snapshot": {
        "type": "object",
        "properties": {
          "bytesRead": {
            "type": "string",
            "format": "uint64"
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/candace.brainspine.v1.ResearchEvent"
            }
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "truncated": {
            "type": "boolean"
          }
        }
      },
...
      "google.rpc.Status": {
        "type": "object",
        "description": "The Status type defines a logical error model suitable for different programming environments.",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "The status code, which should be an enum value of google.rpc.Code."
          },
          "details": {
            "type": "array",
            "description": "A list of messages that carry the error details.",
            "items": {
              "type": "object",
              "properties": {
                "@type": {
                  "type": "string"
                }
              },
              "additionalProperties": true
            }
          },
          "message": {
            "type": "string",
            "description": "A developer-facing error message."
          }
        }
      }
    }
  },
...
}